Refactor backend API and remove unused variables
This commit refactors the backend API to improve its modularity and maintainability. The API endpoints have been separated into multiple files in the `backend/api` directory, and `APIRouter` is used to group related endpoints together.

This commit also removes unused variables from `backend/main.py` and `src/orchestrator.py`.
